+++ puppeteerServer/puppeteerScripts/mangafire.ts
@@ -125,6 +125,7 @@ export async function getManga(
 
     if (chapterResp.status === 429) {
       let retryDelay = (parseInt(chapterResp.headers.get('retry-after')) || 5) * 1000;
+      page.removeAllListeners();
       await page.close();
 
       console.warn(
@@ -158,9 +159,7 @@ export async function getManga(
 
       specialFetchData = vrf;
 
-      await page
-        .goto('about:blank', { waitUntil: 'domcontentloaded', timeout: 1000 })
-        .catch(() => null);
+      await page.setContent('', { waitUntil: 'domcontentloaded' }).catch(() => null);
 
       if (!specialFetchData) throw new Error('Manga: Unable to get VFR!');
       job.log(logWithTimestamp('Loading Chapter Data'));
@@ -176,7 +175,7 @@ export async function getManga(
 
     chapterData = await chapterResp.json();
 
-    if (config.logging.verboseLogging) console.log(chapterData);
+    if (config.debug.verboseLogging) console.log(chapterData);
 
     await job.log(logWithTimestamp('Finished Fetching Chapter Data. Proccessing!'));
     await job.updateProgress(20);
@@ -195,7 +194,7 @@ export async function getManga(
       })
     ).reverse();
 
-    if (config.logging.verboseLogging) console.log(chapterNumbers);
+    if (config.debug.verboseLogging) console.log(chapterNumbers);
 
     if (chapterNumbers.length <= 0)
       throw new Error('Manga: Issue fetching chapters! Please Contact and Admin!');
@@ -222,7 +221,7 @@ export async function getManga(
 
       if (volumeResp.ok) {
         const volumeData = await volumeResp.json();
-        if (config.logging.verboseLogging) console.log(volumeData);
+        if (config.debug.verboseLogging) console.log(volumeData);
         await page.setContent(volumeData.result, { waitUntil: 'domcontentloaded' });
 
         let photoUrls =
@@ -246,7 +245,7 @@ export async function getManga(
 
         photoUrls.reverse();
 
-        if (config.logging.verboseLogging) console.log(photoUrls);
+        if (config.debug.verboseLogging) console.log(photoUrls);
 
         let startingPoint =
           config.updateSettings.refetchImgs || icon ? 0 : Math.max(0, coverIndexes?.length - 1);
@@ -290,10 +289,9 @@ export async function getManga(
       }
     }
     await job.updateProgress(90);
-    await page.close();
     job.log(logWithTimestamp('All Data Fetched processing now'));
 
-    if (config.logging.verboseLogging) {
+    if (config.debug.verboseLogging) {
       console.log(url.split('chapter-').at(-1));
       console.log(chapterNumbers);
     }
@@ -318,11 +316,15 @@ export async function getManga(
   } catch (err) {
     job.log(logWithTimestamp(`Error: ${err}`));
     console.warn('Unable to fetch data for: ' + url);
-    if (config.logging.verboseLogging) console.warn(err);
-    if (!page.isClosed()) await page.close();
+    if (config.debug.verboseLogging) console.warn(err);
 
     //ensure only custom error messages gets sent to user
     if (err.message.startsWith('Manga:')) throw new Error(err.message);
     throw new Error('Unable to fetch Data! maybe invalid Url?');
+  } finally {
+    if (page && !page.isClosed()) {
+      page.removeAllListeners();
+      await page.close().catch(() => {});
+    }
   }
 }