+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # **Change Logs**
 
+## Change Log 3.1.9
+ - Fixed mobile changelog poup size
+ - Resolved MangaPark pages being left open
+ - Puppeteer headless browser config option
+
 ## Change Log 3.1.8
  - Updated all puppeteer server packages.
  - config option for browser path